% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.cerqueira.test.R
\name{graph.cerqueira.test}
\alias{graph.cerqueira.test}
\title{Cerqueira's hypothesis testing for random graphs.}
\usage{
graph.cerqueira.test(
  Graphs1,
  Graphs2,
  maxPer = 300,
  alpha = 0.05,
  printResult = FALSE
)
}
\arguments{
\item{Graphs1}{a list of undirected graphs.
If each graph has the  attribute \code{eigenvalues} containing its
eigenvalues , such values will be used to
compute their spectral density.}

\item{Graphs2}{a list of undirected graphs.
If each graph has the  attribute \code{eigenvalues} containing its
eigenvalues , such values will be used to
compute their spectral density.}

\item{maxPer}{integer indicating the number of bootstrap resamples (default
is \code{300}).}

\item{alpha}{the significance level for the test (default is \code{0.05}).}

\item{printResult}{logical indicating if the test must print the result
(default is \code{FALSE}).}
}
\value{
A list with class "htest" containing the following components:
\itemize{
\item{\code{statistic:}}{ the W-value of the test.}
\item{\code{p.value:}}{ the p-value of the test.}
\item{\code{method:}}{ a string indicating the used method.}
\item{\code{data.name:}}{ a string with the data's name(s).}
}
}
\description{
Andressa Cerqueira, Daniel Fraiman, Claudia D. Vargas and Florencia Leonardi
non-parametric test of hypotheses to verify if two samples of random graphs
were originated from the same probability distribution.
}
\details{
Given two identically independently distributed (idd) samples of graphs \code{Graphs1} and
\code{Graphs2}, the test verifies if they have the same distribution by calculating the
mean distance D from \code{Graphs1} to \code{Graphs2}. The test rejects the null hypothesis if D is
greater than the \code{(1-alpha)}-quantile of the distribution of the test under the
null hypothesis.
}
\examples{
\donttest{
require(igraph)
set.seed(1)

## test under H0
Graphs1 <- Graphs2 <- list()
for(i in 1:10){
  Graphs1[[i]] <- igraph::erdos.renyi.game(50,0.5)
  Graphs2[[i]] <- igraph::erdos.renyi.game(50,0.5)
}
k <- graph.cerqueira.test(Graphs1, Graphs2, printResult = TRUE)

## test under H1
Graphs1 <- Graphs2 <- list()
for(i in 1:10){
  Graphs1[[i]] <- igraph::erdos.renyi.game(50,0.5)
  Graphs2[[i]] <- igraph::erdos.renyi.game(50,0.6)
}
k <- graph.cerqueira.test(Graphs1, Graphs2, printResult = TRUE)
}

}
\references{
Andressa Cerqueira, Daniel Fraiman, Claudia D. Vargas and Florencia Leonardi.
"A test of hypotheses for random graph distributions built from EEG data",
https://ieeexplore.ieee.org/document/7862892
}
