% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.dist.R
\name{graph.dist}
\alias{graph.dist}
\title{Distance matrix on a list of graphs.}
\usage{
graph.dist(Graphs, dist = "JS", ...)
}
\arguments{
\item{Graphs}{a list of undirected graphs.
If each graph has the  attribute \code{eigenvalues} containing its
eigenvalues , such values will be used to
compute their spectral density.}

\item{dist}{string indicating if you want to use the "JS" (default), "L1" or "L2"
distances. "JS" means Jensen-Shannon divergence.}

\item{...}{Other relevant parameters for \code{\link{graph.spectral.density}}.}
}
\value{
a distance matrix
}
\description{
Given a list of graphs, \code{graph.dist} builds a distance matrix
according to the Jensen-Shannon divergence, L2 norm, or L1 norm
between the spectral density of the graphs graphs.
}
\examples{
set.seed(1)
G <- list()
for (i in 1:5) {
  G[[i]] <- igraph::sample_gnp(50, 0.5)
}
graph.dist(G, 3)

}
\keyword{distance_matrix}
