% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.fraiman.test.R
\name{graph.fraiman.test}
\alias{graph.fraiman.test}
\title{Fraiman's hypothesis testing for random graphs.}
\usage{
graph.fraiman.test(Graphs, maxPer = 300, alpha = 0.05, printResult = FALSE)
}
\arguments{
\item{Graphs}{a list of undirected graphs.
If each graph has the  attribute \code{eigenvalues} containing its
eigenvalues , such values will be used to
compute their spectral density.}

\item{maxPer}{integer indicating the number of bootstrap resamples
(default is \code{300}).}

\item{alpha}{the significance level for the test (default is \code{0.05}).}

\item{printResult}{logical indicating if the test must print the result
(default is \code{FALSE}).}
}
\value{
A list with class "htest" containing the following components:
\itemize{
\item{\code{statistic:}}{ the T-value of the test.}
\item{\code{p.value:}}{ the p-value of the test.}
\item{\code{method:}}{ a string indicating the used method.}
\item{\code{data.name:}}{ a string with the data's name(s).}
}
}
\description{
Daniel Fraiman and Ricardo Fraiman test for network differences between
groups with an analysis of variance test (ANOVA).
}
\details{
Given a list of graphs, the test verifies if all the subpopulations have the
same mean network,
under the alternative that at least one subpopulation has a different mean
network.
}
\examples{
\donttest{
set.seed(1)

## test under H0
a <- b <- list()
for(i in 1:10){
  a[[i]] <- igraph::erdos.renyi.game(50,0.5)
  b[[i]] <- igraph::erdos.renyi.game(50,0.5)
}
Graphs <- list(a,b)
k <- graph.fraiman.test(Graphs, printResult = TRUE)

## test under H1
a <- b <- list()
for(i in 1:10){
  a[[i]] <- igraph::erdos.renyi.game(50,0.5)
  b[[i]] <- igraph::erdos.renyi.game(50,0.6)
}
Graphs <- list(a,b)
k <- graph.fraiman.test(Graphs, printResult = TRUE)
}

}
\references{
Fraiman, Daniel, and Ricardo Fraiman. "An ANOVA approach for statistical
comparisons of brain networks",
https://www.nature.com/articles/s41598-018-23152-5
}
