% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.ghoshdastidar.test.R
\name{graph.ghoshdastidar.test}
\alias{graph.ghoshdastidar.test}
\title{Ghoshdastidar hypothesis testing for large random graphs.}
\usage{
graph.ghoshdastidar.test(
  Graphs1,
  Graphs2,
  maxBoot = 300,
  alpha = 0.05,
  two.sample = FALSE,
  printResult = FALSE
)
}
\arguments{
\item{Graphs1}{a list of undirected Graphs.
If each graph has the  attribute \code{eigenvalues} containing its
eigenvalues , such values will be used to
compute their spectral density.}

\item{Graphs2}{a list of undirected Graphs.
If each graph has the  attribute \code{eigenvalues} containing its
eigenvalues , such values will be used to
compute their spectral density.}

\item{maxBoot}{integer indicating the number of bootstrap resamples (default
is \code{300}).}

\item{alpha}{the significance level for the test (default is \code{0.05}).}

\item{two.sample}{logical. If TRUE the sets contain only one graph each. If
FALSE the sets contain more than one graph each (default is \code{FALSE}).}

\item{printResult}{logical indicating if the test must print the result
(default is \code{FALSE}).}
}
\value{
A list with class "htest" containing the following components:
\itemize{
\item{\code{statistic:}}{ the T-value of the test.}
\item{\code{p.value:}}{ the p-value of the test (only returned when the parameter
'two.sample' is \code{FALSE}).}
\item{\code{method:}}{ a string indicating the used method.}
\item{\code{data.name:}}{ a string with the data's name(s).}
}
}
\description{
Given two lists of graphs generated by the inhomogeneous random graph model,
\code{ghoshdastidar} tests if they were generated by the same parameters.
}
\examples{
\donttest{
set.seed(1)

## test for sets with only one graph each under H0
Graphs1 <- Graphs2 <- list()
Graphs1[[1]] <- igraph::erdos.renyi.game(300, 0.6)
Graphs2[[1]] <- igraph::erdos.renyi.game(300, 0.6)
D <- graph.ghoshdastidar.test(Graphs1, Graphs2, two.sample= TRUE, printResult = TRUE)

## test for sets with only one graph each under H1
Graphs1 <- Graphs2 <- list()
Graphs1[[1]] <- igraph::erdos.renyi.game(300, 0.6)
Graphs2[[1]] <- igraph::erdos.renyi.game(300, 0.7)
D <- graph.ghoshdastidar.test(Graphs1, Graphs2, two.sample= TRUE, printResult = TRUE)
}

}
\references{
Ghoshdastidar, Debarghya, et al. "Two-sample tests for large random graphs
using network statistics". arXiv preprint arXiv:1705.06168 (2017).

Ghoshdastidar, Debarghya, et al. "Two-sample hypothesis testing for
inhomogeneous random graphs". arXiv preprint, arXiv:1707.00833 (2017).
}
