% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.tang.test.R
\name{graph.tang.test}
\alias{graph.tang.test}
\title{Tang hypothesis testing for random graphs.}
\usage{
graph.tang.test(
  Graph1,
  Graph2,
  dim,
  sigma = NULL,
  alpha = 0.05,
  bootstrap_sample = 200,
  printResult = FALSE
)
}
\arguments{
\item{Graph1}{the first undirected graph (igraph object).
If \code{Graph} has the  attribute \code{eigenvalues} containing
the eigenvalues of \code{Graph}, such values will be used to
compute its spectral density.}

\item{Graph2}{the second undirected graph (igraph object).
If \code{Graph} has the  attribute \code{eigenvalues} containing
the eigenvalues of \code{Graph}, such values will be used to
compute its spectral density.}

\item{dim}{dimension of the adjacency spectral embedding.}

\item{sigma}{a real value indicating the kernel bandwidth. If \code{NULL} (default)
the bandwidth is calculated by the method.}

\item{alpha}{the significance level for the test (default is \code{0.05}).}

\item{bootstrap_sample}{integer indicating the number of bootstrap resamples
(default is \code{200}).}

\item{printResult}{logical indicating if the test must print the result
(default is \code{FALSE}).}
}
\value{
A list with class "htest" containing the following components:
\itemize{
\item{\code{statistic:}}{ the T-value of the test.}
\item{\code{p.value:}}{ the p-value of the test.}
\item{\code{method:}}{ a string indicating the used method.}
\item{\code{data.name:}}{ a string with the data's name(s).}
}
}
\description{
Given two independent finite-dimensional random dot product graphs,
\code{tang.test} tests if they have generating latent positions that are drawn
from the same distribution.
}
\examples{
set.seed(1)

## test under H0
lpvs <- matrix(rnorm(200), 20, 10)
lpvs <- apply(lpvs, 2, function(x) { return (abs(x)/sqrt(sum(x^2))) })
Graph1 <- igraph::sample_dot_product(lpvs)
Graph2 <- igraph::sample_dot_product(lpvs)
D <- graph.tang.test(Graph1,Graph2, 5, printResult = TRUE)

## test under H1
lpvs2 <- matrix(pnorm(200), 20, 10)
lpvs2 <- apply(lpvs2, 2, function(x) { return (abs(x)/sqrt(sum(x^2))) })
g2 <- suppressWarnings(igraph::sample_dot_product(lpvs2))
D <- graph.tang.test(Graph1,Graph2, 5, printResult = TRUE)


}
\references{
Tang, Minh, et al. "A nonparametric two-sample hypothesis testing problem for
random graphs." Bernoulli 23.3 (2017): 1599-1630.

Tang, Minh, et al. "A semiparametric two-sample hypothesis testing problem
for random graphs." Journal of Computational and Graphical Statistics 26.2
(2017): 344-354.
}
