% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statGraph-package.R
\docType{package}
\name{statGraph-package}
\alias{statGraph}
\alias{statGraph-package}
\title{\packageTitle{statGraph}}
\description{
\packageDescription{statGraph}
}
\details{
The DESCRIPTION file:
\tabular{ll}{
Package: \tab statGraph\cr
Type: \tab Package\cr
Version: \tab 0.4.1\cr
Date: \tab 2020-12-16\cr
Depends: \tab R (>= 3.6.0), stats, graphics\cr
Imports: \tab igraph, MASS, rARPACK, cluster, foreach, parallel, doParallel\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://www.ime.usp.br/~fujita/software.html\cr
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.ime.usp.br/~fujita/software.html}
}

}
\author{
\packageAuthor{statGraph}

Maintainer: \packageMaintainer{statGraph}
}
