% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statGraph-package.R
\docType{package}
\name{statGraph-package}
\alias{statGraph-package}
\title{The statGraph Package}
\description{
Statistical Methods for Graphs.
}
\details{
\tabular{ll}{
Package: \tab statGraph\cr
Type: \tab Package\cr
Version: \tab 0.1.0\cr
Date: \tab 2017-04-21\cr
Depends: \tab R (>= 2.10.0), stats, graphics\cr
Encoding: \tab UTF-8\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
URL: \tab https://www.ime.usp.br/~fujita/software.html\cr
}

Contains statistical methods to analyze graphs, such as graph parameter estimation, model
selection based on the GIC (Graph Information Criterion), statistical tests to
discriminate two or more populations of graphs (ANOGVA -Analysis of Graph Variability), correlation
between graphs, and clustering of graphs.
}
\author{
Suzana S. Santos, Andre Fujita
}
\keyword{package}

